Root password: Make dismissing the report less confusing
The report can't check if the password is set, and anyway,
the user doesn't "have to" set a root password. We only want to
show info here and let the user understand it.

Replace "ignore this problem" with "I understand" and mark it as a suggested_action
button. This essentially does the same, but it makes the experience
less confusing. The action is now affirmative/positive.