Remove root password check
There are pros and cons associated with setting or not
setting a root password. It's difficult to tackle this
topic without confusing users.

To properly cover this topic we need something that isn't
ready yet.. a complete security guide.

In the meantime this test was confusing because:

- Even when the password is set, mintreport had no access
to /etc/shadow and couldn't dismiss the message automatically.
- To simplify the rationale, the message suggested setting a
password, which was a controversial move since it didn't secure
local access anyway and unlocked the root account.

Whatever opinion people may have about this we noticed that, as it is,
it creates far more confusion than guidance with novice users.