Remove correct ownership report
Systemd uses /run for the user runtime dir and none for root. This results in dconf settings for apps running with root privileges being owned by root in ~.

We make it easy for users to run nemo, xed, etc.. with elevated privs and this is the way systemd works right now so we can't treat the presence of root owned files in ~'s dconf cache as something that is not normal.

It's certainly not ideal and it should be fixed, but it's not something we need to flag to the user.